# move validation in distributor into separate middleware