--- a/pkg/distributor/distributor.go
+++ b/pkg/distributor/distributor.go
@@ -806,6 +806,9 @@ func (d *Distributor) prePushRelabelMiddleware(next push.Func) push.Func {
 		}
 
 		if len(removeTsIndexes) > 0 {
+			for _, removeTsIndex := range removeTsIndexes {
+				mimirpb.ReuseTimeseries(req.Timeseries[removeTsIndex].TimeSeries)
+			}
 			req.Timeseries = util.RemoveSliceIndexes(req.Timeseries, removeTsIndexes)
 		}
 
@@ -893,6 +896,9 @@ func (d *Distributor) prePushValidationMiddleware(next push.Func) push.Func {
 			validatedExemplars += len(ts.Exemplars)
 		}
 		if len(removeIndexes) > 0 {
+			for _, removeIndex := range removeIndexes {
+				mimirpb.ReuseTimeseries(req.Timeseries[removeIndex].TimeSeries)
+			}
 			req.Timeseries = util.RemoveSliceIndexes(req.Timeseries, removeIndexes)
 			removeIndexes = removeIndexes[:0]
 		}