--- a/pkg/distributor/distributor.go
+++ b/pkg/distributor/distributor.go
@@ -807,7 +807,7 @@ func (d *Distributor) prePushRelabelMiddleware(next push.Func) push.Func {
 
 		if len(removeTsIndexes) > 0 {
 			for _, removeTsIndex := range removeTsIndexes {
-				mimirpb.ReuseTimeseries(req.Timeseries[removeTsIndex].TimeSeries)
+				mimirpb.ReusePreallocTimeseries(req.Timeseries[removeTsIndex])
 			}
 			req.Timeseries = util.RemoveSliceIndexes(req.Timeseries, removeTsIndexes)
 		}
@@ -897,7 +897,7 @@ func (d *Distributor) prePushValidationMiddleware(next push.Func) push.Func {
 		}
 		if len(removeIndexes) > 0 {
 			for _, removeIndex := range removeIndexes {
-				mimirpb.ReuseTimeseries(req.Timeseries[removeIndex].TimeSeries)
+				mimirpb.ReusePreallocTimeseries(req.Timeseries[removeIndex])
 			}
 			req.Timeseries = util.RemoveSliceIndexes(req.Timeseries, removeIndexes)
 			removeIndexes = removeIndexes[:0]