This change moves the validation of incoming series before the Distributor's forwarding functionality, so that we don't forward invalid series.