@replay 2022-11-09
I just ran the benchmark `BenchmarkDistributor_Push` and I compared the results of `main` against this branch. I'm concerned about the increase in allocations, I'll try to optimize that:

```
name                                                      old time/op    new time/op    delta
Distributor_Push/too_many_labels_limit_reached-10            946µs ± 0%     892µs ± 0%     -5.65%
Distributor_Push/max_label_name_length_limit_reached-10     5.33ms ± 0%    2.95ms ± 0%    -44.69%
Distributor_Push/max_label_value_length_limit_reached-10    3.33ms ± 0%    0.62ms ± 0%    -81.52%
Distributor_Push/timestamp_too_new-10                        579µs ± 0%     636µs ± 0%     +9.95%
Distributor_Push/all_samples_successfully_pushed-10          762µs ± 0%     696µs ± 0%     -8.56%
Distributor_Push/ingestion_rate_limit_reached-10             486µs ± 0%     328µs ± 0%    -32.53%

name                                                      old alloc/op   new alloc/op   delta
Distributor_Push/too_many_labels_limit_reached-10           79.6kB ± 0%  2293.3kB ± 0%  +2780.62%
Distributor_Push/max_label_name_length_limit_reached-10      119kB ± 0%    1047kB ± 0%   +782.56%
Distributor_Push/max_label_value_length_limit_reached-10     104kB ± 0%    1034kB ± 0%   +896.58%
Distributor_Push/timestamp_too_new-10                       88.1kB ± 0%  1021.0kB ± 0%  +1059.20%
Distributor_Push/all_samples_successfully_pushed-10          143kB ± 0%     109kB ± 0%    -23.53%
Distributor_Push/ingestion_rate_limit_reached-10            25.5kB ± 0%     4.7kB ± 0%    -81.38%

name                                                      old allocs/op  new allocs/op  delta
Distributor_Push/too_many_labels_limit_reached-10            2.14k ± 0%     7.15k ± 0%   +233.96%
Distributor_Push/max_label_name_length_limit_reached-10      2.10k ± 0%     6.08k ± 0%   +190.31%
Distributor_Push/max_label_value_length_limit_reached-10     2.10k ± 0%     6.09k ± 0%   +190.79%
Distributor_Push/timestamp_too_new-10                        2.04k ± 0%     6.05k ± 0%   +196.57%
Distributor_Push/all_samples_successfully_pushed-10           48.0 ± 0%      47.0 ± 0%     -2.08%
Distributor_Push/ingestion_rate_limit_reached-10              49.0 ± 0%      44.0 ± 0%    -10.20%
```