@replay 2022-11-09
I know now why the benchmarks show much more allocations in this branch:
Because the new validation middleware shortens the `.Timeseries` and `.Metadata` slices when there are invalid entries and then those entries that get dropped don't get returned to the pools. Will fix that.