@replay 2022-11-09
The commit d84ed7690 fixes this in the new middleware `prePushValidationMiddleware` and also in the old middleware `prePushRelabelMiddleware`, the benchmarks are looking much better now.

```
name                                                      old time/op    new time/op    delta
Distributor_Push/too_many_labels_limit_reached-10            946µs ± 0%     397µs ± 0%  -58.02%
Distributor_Push/max_label_name_length_limit_reached-10     5.33ms ± 0%    2.65ms ± 0%  -50.28%
Distributor_Push/max_label_value_length_limit_reached-10    3.33ms ± 0%    0.41ms ± 0%  -87.82%
Distributor_Push/timestamp_too_new-10                        579µs ± 0%     385µs ± 0%  -33.43%
Distributor_Push/all_samples_successfully_pushed-10          762µs ± 0%     675µs ± 0%  -11.38%
Distributor_Push/ingestion_rate_limit_reached-10             486µs ± 0%     316µs ± 0%  -35.07%

name                                                      old alloc/op   new alloc/op   delta
Distributor_Push/too_many_labels_limit_reached-10           79.6kB ± 0%    84.3kB ± 0%   +5.85%
Distributor_Push/max_label_name_length_limit_reached-10      119kB ± 0%     121kB ± 0%   +2.19%
Distributor_Push/max_label_value_length_limit_reached-10     104kB ± 0%     106kB ± 0%   +2.30%
Distributor_Push/timestamp_too_new-10                       88.1kB ± 0%    92.8kB ± 0%   +5.38%
Distributor_Push/all_samples_successfully_pushed-10          143kB ± 0%     109kB ± 0%  -23.93%
Distributor_Push/ingestion_rate_limit_reached-10            25.5kB ± 0%     4.7kB ± 0%  -81.37%

name                                                      old allocs/op  new allocs/op  delta
Distributor_Push/too_many_labels_limit_reached-10            2.14k ± 0%     2.15k ± 0%   +0.33%
Distributor_Push/max_label_name_length_limit_reached-10      2.10k ± 0%     2.09k ± 0%   -0.14%
Distributor_Push/max_label_value_length_limit_reached-10     2.10k ± 0%     2.09k ± 0%   -0.14%
Distributor_Push/timestamp_too_new-10                        2.04k ± 0%     2.05k ± 0%   +0.34%
Distributor_Push/all_samples_successfully_pushed-10           48.0 ± 0%      45.0 ± 0%   -6.25%
Distributor_Push/ingestion_rate_limit_reached-10              49.0 ± 0%      44.0 ± 0%  -10.20%
```