@replay 2022-11-10
> PR looks fine, but I think it needs a test showing that invalid series are no longer forwarded. Test should fail on main branch, but will pass here. WDYT?

That's a good idea. I added this test: https://github.com/grafana/mimir/pull/3386/files#diff-16b7662ec0e247fa72bf86b63ec8558a8cea51ac2280baf05a25887578132c44R3203-R3331

I force-pushed in order to move the test into the beginning of the PR as the first commit, this allows us to verify that it fails without any further changes by just checking it out directly, when checking out the last commit of the PR it passes:

```
replay@Mauros-MBP mimir % git checkout 7d32a2d125a2f4e5397ad035e1a1b729795f838f
...
HEAD is now at 7d32a2d12 add test to show that validation happens before forwarding
replay@Mauros-MBP mimir % go test -run TestValidationBeforeForwarding github.com/grafana/mimir/pkg/distributor 
--- FAIL: TestValidationBeforeForwarding (1.00s)
    distributor_test.go:3318: 
                Error Trace:    /Users/replay/go/src/github.com/grafana/mimir/pkg/distributor/distributor_test.go:3318
                Error:          An error is expected but got nil.
                Test:           TestValidationBeforeForwarding
    distributor_test.go:3320: 
                Error Trace:    /Users/replay/go/src/github.com/grafana/mimir/pkg/distributor/distributor_test.go:3320
                Error:          Should be true
                Test:           TestValidationBeforeForwarding
FAIL
FAIL    github.com/grafana/mimir/pkg/distributor        2.352s
FAIL
replay@Mauros-MBP mimir % git checkout 193c8f220943bc6c11df9518eb98a02d375825de
Previous HEAD position was 7d32a2d12 add test to show that validation happens before forwarding
HEAD is now at 193c8f220 return timeseries to pool
replay@Mauros-MBP mimir % go test -run TestValidationBeforeForwarding github.com/grafana/mimir/pkg/distributor
ok      github.com/grafana/mimir/pkg/distributor        1.506s
```